requirements.txt
# Core Dependencies - Updated to latest compatible versions
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
langchain-core==0.3.29
langchain-text-splitters==0.3.5

# OpenAI - Latest stable
openai==1.58.1

# Knowledge Graph - Updated to latest
graphiti-core==0.3.6
neo4j==5.27.0

# Vector Store
faiss-cpu==1.9.0.post1
tiktoken==0.8.0

# Utilities
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.3

# Visualization
matplotlib==3.10.0
pyvis==0.3.2
networkx==3.4.2

# Progress and CLI
tqdm==4.67.1
rich==13.9.4

# Additional dependencies
pydantic==2.10.5
pydantic-settings==2.7.1
